backend/requirements.txt
# ==============================================================================
# CORE DJANGO DEPENDENCIES
# ==============================================================================
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# ==============================================================================
# DATABASE & ORM
# ==============================================================================
psycopg2-binary==2.9.9
django-model-utils==4.3.1
django-bulk-update==2.2.0

# ==============================================================================
# AUTHENTICATION & AUTHORIZATION
# ==============================================================================
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
django-oauth-toolkit==1.7.1
django-guardian==2.4.0

# ==============================================================================
# CACHING & PERFORMANCE
# ==============================================================================
redis==5.0.1
django-redis==5.4.0
django-cachalot==2.6.1
django-silk==5.0.4

# ==============================================================================
# API & SERIALIZATION
# ==============================================================================
djangorestframework-camel-case==1.4.2
django-filter==23.4
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# ==============================================================================
# REAL-TIME & WEBSOCKETS
# ==============================================================================
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# ==============================================================================
# CELERY & BACKGROUND TASKS
# ==============================================================================
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
kombu==5.3.4

# ==============================================================================
# FILE HANDLING & STORAGE
# ==============================================================================
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
botocore==1.34.0

# ==============================================================================
# VALIDATION & UTILITIES
# ==============================================================================
python-decouple==3.8
python-slugify==8.0.1
phonenumbers==8.13.26
django-phonenumber-field==7.2.0
django-countries==7.5.1

# ==============================================================================
# MONITORING & LOGGING
# ==============================================================================
sentry-sdk==1.38.0
django-health-check==3.17.0
structlog==23.2.0
django-structlog==6.0.0

# ==============================================================================
# DEVELOPMENT & DEBUGGING
# ==============================================================================
django-debug-toolbar==4.2.0
django-querycount==0.8.3
factory-boy==3.3.0
faker==20.1.0

# ==============================================================================
# TESTING
# ==============================================================================
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
model-bakery==1.17.0

# ==============================================================================
# CODE QUALITY
# ==============================================================================
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
django-stubs==4.2.7
djangorestframework-stubs==3.14.5

# ==============================================================================
# SECURITY
# ==============================================================================
django-ratelimit==4.1.0
django-axes==6.1.1
cryptography==41.0.7

# ==============================================================================
# EMAIL & NOTIFICATIONS
# ==============================================================================
django-anymail==10.2
sendgrid==6.10.0

# ==============================================================================
# DEPLOYMENT & PRODUCTION
# ==============================================================================
gunicorn==21.2.0
whitenoise==6.6.0
django-compressor==4.4

# ==============================================================================
# UTILITIES
# ==============================================================================
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30
